Addressing Subhankar's comments
