+++ examples/nlp/language_modeling/conf/t5_speechlm_pretraining.yaml
@@ -54,6 +54,7 @@ model:
   model_seq_length: 2048
   seq_pattern: "parallel" # parallel, delay_parallel, flatten
   speech_head_type: "token_level" # token_level, linear
+  cross_entropy_type: "regular" # regular, vocab_parallel
   restore_path: null # Path to an existing p-tuned/prompt tuned .nemo model you wish to add new tasks to or run inference with
   language_model_path: ??? # Path to the pretrained T5 language model .nemo file, always required
   save_nemo_on_validation_end: True # Saves an inference ready .nemo file every time a checkpoint is saved during training.
