+++ examples/nlp/language_modeling/megatron_t5_speechlm.py
@@ -79,7 +79,7 @@ def main(cfg) -> None:
         cfg.model.precision = cfg.trainer.precision
 
     # load existing or init new soft prompt T5 model
-    if cfg.model.get("restore_path", None) is not None and cfg.model.restore_path.endswith(".nemo"):
+    if cfg.model.get("restore_path", None) is not None:
         print(f"cfg.model.restore_path {cfg.model.restore_path}")
         model = MegatronT5SpeechLMModel.restore_from(
             cfg.model.restore_path, cfg.model, trainer=trainer, save_restore_connector=NLPSaveRestoreConnector()
@@ -88,11 +88,7 @@ def main(cfg) -> None:
     else:
         print(f"cfg.model.restore_path is None")
         model = MegatronT5SpeechLMModel(cfg.model, trainer=trainer)
-        if cfg.model.get("restore_path", None) is not None:
-            print("Loading state dict")
-            ckpt = torch.load(cfg.model.restore_path, map_location='cpu')
-            model.load_state_dict(ckpt['state_dict'], strict=False)
-            print("Loaded state dict")
+        model.maybe_init_from_pretrained_checkpoint(cfg=cfg)
 
     trainer.fit(model)
