+++ examples/nlp/language_modeling/speechlm_t5_inference.py
@@ -71,20 +71,6 @@ def main(cfg) -> None:
     with open_dict(cfg):
         cfg.model.precision = cfg.trainer.precision
 
-    # checkpoint_path = "/home/shehzeenh/Code/NeMo/nemo_experiments/p_tuning_squad_t5/checkpoints/Step34770.ckpt"
-    # checkpoint_path = "/Data/Experiments/full_data_tts/p_tuning_squad_t5/checkpoints/Step46794.ckpt"
-    # checkpoint_path = "/Data/Experiments/full_data_tts/p_tuning_squad_t5/checkpoints/Step100k.ckpt"
-    # checkpoint_path = "/Data/Experiments/NewDatasetClassOverfitting1000/p_tuning_squad_t5/checkpoints/Step164164.ckpt"
-    # checkpoint_path = "/Data/Experiments/Selene_Runs/Run512_24000_AdamW-220M-unfreeze-tts_lr5e-5_nosched_posemb512_FP32/p_tuning_squad_t5/2023-08-14_19-49-38/checkpoints/Step32697.ckpt"
-    # checkpoint_path = "/Data/Experiments/FromPretrain30k.ckpt"
-    # checkpoint_path = "/Data/Experiments/breakCode_30ktrainingExamples_512/p_tuning_squad_t5/checkpoints/Step38043.ckpt"
-    # checkpoint_path = "/Data/Experiments/breakCode_30ktrainingExamples_512/Step16k.ckpt"
-    # checkpoint_path = "/Data/Experiments/Selene_Runs/Run512_24000_AdamW-220M-unfreeze-tts_lr5e-5_nosched_posemb512_FP32/p_tuning_squad_t5/2023-08-14_19-49-38/checkpoints/Step47229.ckpt"
-    # checkpoint_path = "/Data/Experiments/Selene_Runs/NewDL_24000_AdamW-220M-unfreeze-tts_lr5e-5_nosched_posemb1536_FP32/p_tuning_squad_t5/2023-08-14_19-49-38_latest/checkpoints/Step89200.ckpt"
-    # checkpoint_path = "/Data/Experiments/ComparisonBreakUnbreak/NoBreak/p_tuning_squad_t5/checkpoints/Step29k.ckpt"
-    # checkpoint_path = "/Data/Experiments/ComparisonBreakUnbreak/WithBreak/p_tuning_squad_t5/checkpoints/Step82348.ckpt"
-    # checkpoint_path = "/Data/Experiments/ComparisonBreakUnbreak/NoBreak/p_tuning_squad_t5/checkpoints/Step82k.ckpt"
-    # checkpoint_path = "/Data/Experiments/ComparisonBreakUnbreak/FromPretrainedDelayParallelLinearHead/Epoch10.ckpt"
     checkpoint_path = "/Data/Experiments/ComparisonBreakUnbreak/DelayParallelLinearHead/Epoch10.ckpt"
     model = MegatronT5SpeechLMModel.load_from_checkpoint(
         checkpoint_path=checkpoint_path, trainer=trainer, cfg=cfg.model
